+++ snapshot-tests-core/src/main/java/de/skuzzle/test/snapshots/impl/SnapshotConfiguration.java
@@ -14,7 +14,29 @@ import org.apiguardian.api.API.Status;
  * @see DefaultSnapshotConfiguration
  */
 @API(status = Status.INTERNAL, since = "1.1.0")
-interface SnapshotConfiguration {
+public interface SnapshotConfiguration {
+
+    /**
+     * Creates a SnapshotConfiguration for the given test class.
+     *
+     * @param testClass The test class.
+     * @return The configuration.
+     * @since 1.7.0
+     */
+    @API(status = Status.INTERNAL, since = "1.7.0")
+    static SnapshotConfiguration defaultConfigurationFor(Class<?> testClass) {
+        return DefaultSnapshotConfiguration.forTestClass(testClass);
+    }
+
+    /**
+     *
+     * @deprecated Since 1.7.0 - Only introduced for backward compatibility.
+     */
+    @Deprecated(since = "1.7.0", forRemoval = true)
+    @API(status = Status.DEPRECATED, since = "1.7.0")
+    static SnapshotConfiguration legacyConfigurationFor(Class<?> testClass) {
+        return LegacySnapshotConfiguration.forTestClass(testClass);
+    }
 
     /**
      * Determines the directory into which the snapshot files for the currently executed
@@ -61,7 +83,11 @@ interface SnapshotConfiguration {
      * results are collected and processed at once when the test method finishes.
      *
      * @return Whether to use soft assertions.
+     * @deprecated Since 1.7.0 - Soft assertion will no longer be supported with version
+     *             2.0
      */
+    @Deprecated(since = "1.7.0", forRemoval = true)
+    @API(status = Status.DEPRECATED, since = "1.7.0")
     boolean isSoftAssertions();
 
 }