+++ snapshot-tests-core/src/test/java/de/skuzzle/test/snapshots/SnapshotsTest.java
@@ -10,6 +10,7 @@ import org.junit.jupiter.params.provider.ValueSource;
 
 import de.skuzzle.test.snapshots.SnapshotDsl.Snapshot;
 import de.skuzzle.test.snapshots.SnapshotTestResult.SnapshotStatus;
+import de.skuzzle.test.snapshots.junit5.EnableSnapshotTests;
 
 @EnableSnapshotTests
 public class SnapshotsTest {