+++ snapshot-tests-html/src/test/java/de/skuzzle/test/snapshots/html/HtmlSnapshotTests.java
@@ -6,10 +6,10 @@ import java.util.regex.Pattern;
 
 import org.junit.jupiter.api.Test;
 
-import de.skuzzle.test.snapshots.EnableSnapshotTests;
 import de.skuzzle.test.snapshots.SnapshotDsl.Snapshot;
 import de.skuzzle.test.snapshots.SnapshotTestResult;
 import de.skuzzle.test.snapshots.data.html.HtmlSnapshot;
+import de.skuzzle.test.snapshots.junit5.EnableSnapshotTests;
 
 @EnableSnapshotTests
 public class HtmlSnapshotTests {